refactor(appstudioplugin): migrate IAppDefinition
