+++ packages/fx-core/tests/plugins/resource/appstudio/unit/validate.test.ts
@@ -10,48 +10,50 @@ import { AppStudioPlugin } from "./../../../../../src/plugins/resource/appstudio
 import { ConfigMap, PluginContext, TeamsAppManifest, ok } from "@microsoft/teamsfx-api";
 
 describe("validate manifest", () => {
-    let plugin: AppStudioPlugin;
-    let ctx: PluginContext;
+  let plugin: AppStudioPlugin;
+  let ctx: PluginContext;
 
-    beforeEach(async () => {
-        plugin = new AppStudioPlugin();
-        ctx = {
-            root: "./",
-            configOfOtherPlugins: new Map(),
-            config: new ConfigMap(),
-            app: new TeamsAppManifest()
-        }
-    })
+  beforeEach(async () => {
+    plugin = new AppStudioPlugin();
+    ctx = {
+      root: "./",
+      configOfOtherPlugins: new Map(),
+      config: new ConfigMap(),
+      app: new TeamsAppManifest(),
+    };
+  });
 
-    it("valid manifest", async () => {
-        const manifestFile = path.resolve(__dirname, "./../resources/valid.manifest.json");
-        const manifest = await fs.readJson(manifestFile);
-        const manifestString = JSON.stringify(manifest);
+  it("valid manifest", async () => {
+    const manifestFile = path.resolve(__dirname, "./../resources/valid.manifest.json");
+    const manifest = await fs.readJson(manifestFile);
+    const manifestString = JSON.stringify(manifest);
 
-        sinon.stub(plugin, "validateManifest").resolves(ok([]));
-        
-        const validationResult = await plugin.validateManifest(ctx, manifestString);
-        chai.assert.isTrue(validationResult.isOk());
-        if (validationResult.isOk()) {
-            chai.expect(validationResult.value).to.have.lengthOf(0);
-        }
+    sinon.stub(plugin, "validateManifest").resolves(ok([]));
 
-        sinon.restore();
-    });
+    const validationResult = await plugin.validateManifest(ctx, manifestString);
+    chai.assert.isTrue(validationResult.isOk());
+    if (validationResult.isOk()) {
+      chai.expect(validationResult.value).to.have.lengthOf(0);
+    }
 
-    it("invalid manifest", async () => {
-        const manifestFile = path.resolve(__dirname, "./../resources/invalid.manifest.json");
-        const manifest = await fs.readJson(manifestFile);
-        const manifestString = JSON.stringify(manifest);
+    sinon.restore();
+  });
 
-        sinon.stub(plugin, "validateManifest").resolves(ok(["developer | Required properties are missing from object: []."]));
+  it("invalid manifest", async () => {
+    const manifestFile = path.resolve(__dirname, "./../resources/invalid.manifest.json");
+    const manifest = await fs.readJson(manifestFile);
+    const manifestString = JSON.stringify(manifest);
 
-        const validationResult = await plugin.validateManifest(ctx, manifestString);
-        chai.assert.isTrue(validationResult.isOk());
-        if (validationResult.isOk()) {
-            chai.expect(validationResult.value).to.have.lengthOf(1);
-        }
+    sinon
+      .stub(plugin, "validateManifest")
+      .resolves(ok(["developer | Required properties are missing from object: []."]));
 
-        sinon.restore();
-    });
+    const validationResult = await plugin.validateManifest(ctx, manifestString);
+    chai.assert.isTrue(validationResult.isOk());
+    if (validationResult.isOk()) {
+      chai.expect(validationResult.value).to.have.lengthOf(1);
+    }
+
+    sinon.restore();
+  });
 });
